+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 ---
 ###5.10_WIP
 - Added -iplocation
+- Added subreddit information
+- Added reddit post information
+- Added reddit comment information
 - Re-added duration to YouTube stats
 - Links within Tweets now get handled, too (YouTube stats, link titles, GitHub stats, etc. now show)
 - Fixed (some?) errors not getting logged to file
-- Fixed SpellingCorrection not working when a user corrects someone else twice in a row
+- Fixed SpellingCorrection not working if a user corrects someone else twice in a row
 - Internal changes
 
 ###5.9.4